Move responses into different crate
