+++ src/server/mod.rs
@@ -1,6 +1,8 @@
 pub mod config;
 #[allow(clippy::module_inception)]
 pub mod server;
+pub mod response;
 
+pub use self::response::*;
 pub use self::config::Config;
 pub use self::server::Server;
